v1.3.2
## Summary
This release contains a small quality of life addition for refactor mode. When refactor mode is enabled and a developer attempts to enter play mode, a modal warning will be shown that it should be disabled before playing. If **Disable Refactor Mode** is selected, play mode will be stopped and refactor mode disabled, which will cause a script recompile. If **Enter Playmode** is selected, playmode will proceed and a warning issued in the console.

**Note:** If you install EntitasRedux using a `.UnityPackage` you must also grab the latest release of Genesis as well from [here](https://github.com/jeffcampbellmakesgames/Genesis) as it is not included. For better dependency handling, I would recommend using OpenUPM instead.

## ADDED
* Added a warning modal dialog to be shown when when entering playmode and refactor mode is enabled.

## CHANGED
* Refactored some internals for refactor mode for easy reuse elsewhere in the framework.

## FIXED
* NA

## REMOVED
* NA